perf(native): emit board planes as uint8, expand on GPU
The 104 board planes were shipped to the GPU as float32 and cast to fp16 on
the CPU, but 103 of them are strictly binary.  The actor now encodes uint8
(Position::encode_planes_u8) and InferenceRuntime expands on-device.

The one non-binary plane is the halfmove clock, which the float encoder
writes as float(clock / 100.0).  The compact encoder stores the RAW clock and
the consumer divides by HALFMOVE_SCALE in FP32 before narrowing.  That
round-trip is BIT-EXACT for every storable clock value: float32(double(n)/100)
== float32(float32(n)/100.0f) for all n in 0..255, asserted exhaustively.
Both the plane index and the scale are exported from the extension
(HALFMOVE_META_PLANE / HALFMOVE_SCALE / HALFMOVE_CLAMP) so the constant cannot
drift from the encoder.

Adds native_selfplay.expand_planes() for host-side consumers standing in for
the network; benchmarks/native_mcts.py::FakeEvaluator now uses it.  That stand-in
reads plane 102 — the halfmove plane — so it had been consuming the raw clock,
a 100x error that broke native-vs-python MCTS parity.  With the expansion the
pre-existing arena determinism golden (GOLDEN_TRANSCRIPT_HASH, recorded before
this change) still matches, which is the evidence that the wire-format change
is behaviour-neutral.

InferenceRuntime accepts float32 and uint8; float32 stays valid.

Measured (RTX 2080 Ti, uncontended):
  prepare() B=256   float32 0.329 ms / 6.82 MB  ->  uint8 0.195 ms / 1.70 MB
  inference phase   1.227 s -> 1.001 s over 250 rounds (-18%)
  end-to-end rounds/s  41.70 -> 42.29 (+1.4%; gather is 77% of the round)

CORRECTION to the earlier review: prepare() was reported at 8.59 ms/round.
That measurement was taken while the live trainer held the GPU; uncontended
it is 0.33 ms.  The compact-plane win is real but roughly an order of
magnitude smaller than that figure implied.

Also documents an upstream behaviour found while testing: the vendored
chess-library truncates the FEN halfmove-clock field modulo 256 while
parsing, so a FEN claiming 300 is 44 to BOTH encoders.  Pre-existing, not a
regression, unreachable in self-play (the 75-move rule ends games at 150).

Tests: tests/test_compact_planes.py (34 new).

+++ tests/test_native_actor.py
@@ -163,7 +163,10 @@ def test_gather_leaves_merges_across_games():
     assert B > 1
     assert tokens == list(range(B))
     assert inputs.shape == (B, PLANES, ROWS, COLS)
-    assert inputs.dtype == np.float32
+    # Compact-plane contract: gather_leaves emits uint8 planes (binary
+    # planes 0/1, halfmove plane = raw clock).  See
+    # tests/test_compact_planes.py for the exact-expansion proof.
+    assert inputs.dtype == np.uint8
     assert inputs.flags["C_CONTIGUOUS"]
     assert offsets.dtype == np.int32
     assert indices.dtype == np.int32
